requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
frozenlist==1.4.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
kiwisolver==1.4.5
magic-filter==1.0.12
matplotlib==3.9.0
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
tenacity==8.3.0
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
